requirements.txt
streamlit
streamlit_option_menu
pandas
streamlit_shadcn_ui
youtube_transcript_api
google-generativeai